#2 Restructured file organization. Also added a Users enum.
