+++ src/battleship/enumerations/Users.java
@@ -1,0 +1,6 @@
+package battleship.enumerations;
+
+public enum Users {
+	PLAYER,
+	COMPUTER
+}
