+++ src/battleship/userInterface/SpaceButton.java
@@ -1,4 +1,4 @@
-package battleship.grid;
+package battleship.userInterface;
 
 import java.awt.Color;
 import java.awt.Graphics;
